+++ neoforge/src/main/kotlin/miragefairy2024/neoforge/MirageFairy2024NeoForgeReiServerPlugin.kt
@@ -5,16 +5,15 @@ import me.shedaniel.rei.api.common.entry.comparison.ItemComparatorRegistry
 import me.shedaniel.rei.api.common.plugins.REIServerPlugin
 import me.shedaniel.rei.forge.REIPluginCommon
 import miragefairy2024.mod.recipeviewer.rei.ReiEvents
-import mirrg.kotlin.helium.Single
 
 @REIPluginCommon
 @Suppress("unused")
 class MirageFairy2024NeoForgeReiServerPlugin : REIServerPlugin {
     override fun registerDisplaySerializer(registry: DisplaySerializerRegistry) {
-        ReiEvents.onRegisterDisplaySerializer.fire { it(Single(registry)) }
+        ReiEvents.onRegisterDisplaySerializer.fire { it(registry) }
     }
 
     override fun registerItemComparators(registry: ItemComparatorRegistry) {
-        ReiEvents.onRegisterItemComparators.fire { it(Single(registry)) }
+        ReiEvents.onRegisterItemComparators.fire { it(registry) }
     }
 }
